@@ -2,7 +2,7 @@
     <Import Project="..\..\DiscordBotCore.props" />
     <PropertyGroup>
         <OutputType>Exe</OutputType>
-        <TargetFramework>netcoreapp2.1</TargetFramework>
+        <TargetFramework>netcoreapp3.0</TargetFramework>
         <RuntimeIdentifiers>$(RuntimeIdentifiers);win10-arm</RuntimeIdentifiers>
     </PropertyGroup>
     <ItemGroup>
