@@ -16,35 +16,35 @@ namespace MechHisui.Superfight.Preconditions
         }
 
         public override Task<PreconditionResult> CheckPermissionsAsync(
-            ICommandContext context,
-            CommandInfo command,
-            IServiceProvider services)
+            ICommandContext context, CommandInfo command, IServiceProvider services)
         {
             var sfservice = services.GetService<SuperfightService>();
-            if (sfservice != null)
-            {
-                var authorId = context.User.Id;
-                var game = sfservice.GetGameFromChannel(context.Channel);
+            if (sfservice is null)
+                return Task.FromResult(PreconditionResult.FromError("No service."));
 
-                if (game != null)
-                {
-                    var fighter1 = game.TurnPlayers[0];
-                    var fighter2 = game.TurnPlayers[1];
+            var authorId = context.User.Id;
+            var game = sfservice.GetGameFromChannel(context.Channel);
 
-                    switch (Role)
-                    {
-                        case PlayerRole.Fighter:
-                            return (fighter1.User.Id == context.User.Id || fighter2.User.Id == context.User.Id)
-                                ? Task.FromResult(PreconditionResult.FromSuccess())
-                                : Task.FromResult(PreconditionResult.FromError("Cannot use command at this time."));
-                        case PlayerRole.NonFighter:
-                            return !(fighter1.User.Id == context.User.Id || fighter2.User.Id == context.User.Id)
-                                ? Task.FromResult(PreconditionResult.FromSuccess())
-                                : Task.FromResult(PreconditionResult.FromError("Cannot use command at this time."));                    }
-                }
+            if (game is null)
                 return Task.FromResult(PreconditionResult.FromError("No game."));
+
+            var fighter1 = game.TurnPlayers[0];
+            var fighter2 = game.TurnPlayers[1];
+
+            switch (Role)
+            {
+                case PlayerRole.Fighter:
+                    return (fighter1.User.Id == context.User.Id || fighter2.User.Id == context.User.Id)
+                        ? Task.FromResult(PreconditionResult.FromSuccess())
+                        : Task.FromResult(PreconditionResult.FromError("Cannot use command at this time."));
+                case PlayerRole.NonFighter:
+                    return !(fighter1.User.Id == context.User.Id || fighter2.User.Id == context.User.Id)
+                        ? Task.FromResult(PreconditionResult.FromSuccess())
+                        : Task.FromResult(PreconditionResult.FromError("Cannot use command at this time."));
+
+                default:
+                    return Task.FromResult(PreconditionResult.FromError("Cannot use command at this time."));
             }
-            return Task.FromResult(PreconditionResult.FromError("No service."));
         }
     }
 
