@@ -1,9 +1,10 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
+#nullable disable warnings
 namespace MechHisui.Core
 {
-    public sealed class NamedScalar
+    public sealed class Variant
     {
         [Key, DatabaseGenerated(DatabaseGeneratedOption.None)]
         public string Key { get; set; }
@@ -11,5 +12,11 @@ namespace MechHisui.Core
         public string StringValue { get; set; }
 
         public int IntValue { get; set; }
+    }
+
+    public enum VariantType
+    {
+        Int,
+        String
     }
 }
