@@ -1,10 +1,9 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
-using System.Linq;
 using MechHisui.FateGOLib;
 
+#nullable disable warnings
 namespace MechHisui.Core
 {
     public sealed class CERange : ICERange
