@@ -1,4 +1,6 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.Extensions.DependencyInjection;
 using Discord.Commands;
@@ -6,15 +8,8 @@ using Discord.Commands;
 namespace MechHisui.ExplodingKittens
 {
     [AttributeUsage(AttributeTargets.Method | AttributeTargets.Class, AllowMultiple = false, Inherited = false)]
-    internal sealed class RequireTurnPlayerAttribute : PreconditionAttribute
+    internal sealed class RequirePlayerAttribute : PreconditionAttribute
     {
-        private readonly bool _equalingTurnPlayer;
-
-        public RequireTurnPlayerAttribute(bool equalingTurnPlayer)
-        {
-            _equalingTurnPlayer = equalingTurnPlayer;
-        }
-
         public override Task<PreconditionResult> CheckPermissionsAsync(
             ICommandContext context, CommandInfo command, IServiceProvider services)
         {
@@ -27,11 +22,9 @@ namespace MechHisui.ExplodingKittens
                 return Task.FromResult(PreconditionResult.FromError("No game active in this channel."));
 
             var authorId = context.User.Id;
-            var turnPlayerId = game.TurnPlayer.Value.User.Id;
-
-            return ((authorId == turnPlayerId) == _equalingTurnPlayer)
+            return (game.PlayerIds.Contains(authorId))
                 ? Task.FromResult(PreconditionResult.FromSuccess())
-                : Task.FromResult(PreconditionResult.FromError("Cannot use command at this time."));
+                : Task.FromResult(PreconditionResult.FromError("Command can only be used by a player."));
         }
     }
 }
