@@ -1,13 +1,14 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using Discord.Addons.MpGame;
 
 namespace MechHisui.SecretHitler
 {
-    public interface ISecretHitlerConfig
+    public interface ISecretHitlerConfig : IMpGameServiceConfig
     {
         Task<IEnumerable<string>> GetThemeKeysAsync();
 
-        Task<ISecretHitlerTheme> GetThemeAsync(string key);
+        Task<ISecretHitlerTheme?> GetThemeAsync(string key);
     }
 }
