@@ -1,132 +1,134 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿//using System;
+//using System.Collections.Generic;
+//using System.Text;
 
-namespace SharedExtensions
-{
-    internal static class SpanExtensions
-    {
-        public static ReadOnlySpan<char> SliceUntil(this ReadOnlySpan<char> span, char delimeter, out ReadOnlySpan<char> remainder)
-        {
-            for (int i = 0; i < span.Length; i++)
-            {
-                if (span[i] == delimeter && i > 0)
-                {
-                    int remStart = i + 1; //skip the delimeter
-                    remainder = span.Slice(remStart, span.Length - remStart).Trim();
-                    return span.Slice(0, i).Trim();
-                }
-            }
-            remainder = ReadOnlySpan<char>.Empty;
-            return span;
-        }
+//namespace SharedExtensions
+//{
+//    internal static class SpanExtensions
+//    {
+//        public static ReadOnlySpan<char> SliceUntil(
+//            this ReadOnlySpan<char> span, char delimeter, out ReadOnlySpan<char> remainder)
+//        {
+//            for (int i = 0; i < span.Length; i++)
+//            {
+//                if (span[i] == delimeter && i > 0)
+//                {
+//                    // add 1 to skip the delimeter itself
+//                    remainder = span[(i + 1)..].Trim();
+//                    return span.Slice(0, i).Trim();
+//                }
+//            }
 
-        public static string Materialize(this ReadOnlySpan<char> span)
-            => new string(span.ToArray());
+//            remainder = ReadOnlySpan<char>.Empty;
+//            return span;
+//        }
 
-        public static ReadOnlySpan<char> TrimBraces(this ReadOnlySpan<char> span)
-            => (CharAliasMap.ContainsKey(span[0]))
-                ? span.Slice(1, FindMatchingBrace(span) - 1)
-                : span;
+//        public static string Materialize(this ReadOnlySpan<char> span)
+//            => new string(span.ToArray());
 
-        public static int FindMatchingBrace(this ReadOnlySpan<char> span)
-        {
-            var braces = new Stack<char>();
+//        public static ReadOnlySpan<char> TrimBraces(this ReadOnlySpan<char> span)
+//            => (CharAliasMap.ContainsKey(span[0]))
+//                ? span[1..FindMatchingBrace(span)]
+//                : span;
 
-            for (int i = 0; i < span.Length; i++)
-            {
-                var current = span[i];
-                var c = (braces.Count > 0)
-                    ? braces.Peek()
-                    : default;
+//        public static int FindMatchingBrace(this ReadOnlySpan<char> span)
+//        {
+//            var braces = new Stack<char>();
 
-                if (c == current)
-                {
-                    braces.Pop();
-                    if (braces.Count == 0)
-                        return i;
-                }
-                else if (CharAliasMap.TryGetValue(current, out var match))
-                {
-                    braces.Push(match);
-                }
-            }
+//            for (int i = 0; i < span.Length; i++)
+//            {
+//                var current = span[i];
+//                var c = (braces.Count > 0)
+//                    ? braces.Peek()
+//                    : default;
 
-            throw new InvalidOperationException();
-        }
+//                if (c == current)
+//                {
+//                    braces.Pop();
+//                    if (braces.Count == 0)
+//                        return i;
+//                }
+//                else if (CharAliasMap.TryGetValue(current, out var match))
+//                {
+//                    braces.Push(match);
+//                }
+//            }
 
-        // Output of a gist provided by https://gist.github.com/ufcpp
-        // https://gist.github.com/ufcpp/5b2cf9a9bf7d0b8743714a0b88f7edc5
-        internal static readonly Dictionary<char, char> CharAliasMap
-            = new Dictionary<char, char> {
-                    {'\"', '\"' },
-                    {'[', ']' },
-                    {'{', '}' },
-                    {'«', '»' },
-                    {'‘', '’' },
-                    {'“', '”' },
-                    {'„', '‟' },
-                    {'‹', '›' },
-                    {'‚', '‛' },
-                    {'《', '》' },
-                    {'〈', '〉' },
-                    {'「', '」' },
-                    {'『', '』' },
-                    {'〝', '〞' },
-                    {'﹁', '﹂' },
-                    {'﹃', '﹄' },
-                    {'＂', '＂' },
-                    {'＇', '＇' },
-                    {'｢', '｣' },
-                    {'(', ')' },
-                    {'༺', '༻' },
-                    {'༼', '༽' },
-                    {'᚛', '᚜' },
-                    {'⁅', '⁆' },
-                    {'⌈', '⌉' },
-                    {'⌊', '⌋' },
-                    {'❨', '❩' },
-                    {'❪', '❫' },
-                    {'❬', '❭' },
-                    {'❮', '❯' },
-                    {'❰', '❱' },
-                    {'❲', '❳' },
-                    {'❴', '❵' },
-                    {'⟅', '⟆' },
-                    {'⟦', '⟧' },
-                    {'⟨', '⟩' },
-                    {'⟪', '⟫' },
-                    {'⟬', '⟭' },
-                    {'⟮', '⟯' },
-                    {'⦃', '⦄' },
-                    {'⦅', '⦆' },
-                    {'⦇', '⦈' },
-                    {'⦉', '⦊' },
-                    {'⦋', '⦌' },
-                    {'⦍', '⦎' },
-                    {'⦏', '⦐' },
-                    {'⦑', '⦒' },
-                    {'⦓', '⦔' },
-                    {'⦕', '⦖' },
-                    {'⦗', '⦘' },
-                    {'⧘', '⧙' },
-                    {'⧚', '⧛' },
-                    {'⧼', '⧽' },
-                    {'⸂', '⸃' },
-                    {'⸄', '⸅' },
-                    {'⸉', '⸊' },
-                    {'⸌', '⸍' },
-                    {'⸜', '⸝' },
-                    {'⸠', '⸡' },
-                    {'⸢', '⸣' },
-                    {'⸤', '⸥' },
-                    {'⸦', '⸧' },
-                    {'⸨', '⸩' },
-                    {'【', '】'},
-                    {'〔', '〕' },
-                    {'〖', '〗' },
-                    {'〘', '〙' },
-                    {'〚', '〛' }
-                };
-    }
-}
+//            throw new InvalidOperationException();
+//        }
+
+//        // Output of a gist provided by https://gist.github.com/ufcpp
+//        // https://gist.github.com/ufcpp/5b2cf9a9bf7d0b8743714a0b88f7edc5
+//        internal static readonly Dictionary<char, char> CharAliasMap
+//            = new Dictionary<char, char> {
+//                    {'\"', '\"' },
+//                    {'[', ']' },
+//                    {'{', '}' },
+//                    {'«', '»' },
+//                    {'‘', '’' },
+//                    {'“', '”' },
+//                    {'„', '‟' },
+//                    {'‹', '›' },
+//                    {'‚', '‛' },
+//                    {'《', '》' },
+//                    {'〈', '〉' },
+//                    {'「', '」' },
+//                    {'『', '』' },
+//                    {'〝', '〞' },
+//                    {'﹁', '﹂' },
+//                    {'﹃', '﹄' },
+//                    {'＂', '＂' },
+//                    {'＇', '＇' },
+//                    {'｢', '｣' },
+//                    {'(', ')' },
+//                    {'༺', '༻' },
+//                    {'༼', '༽' },
+//                    {'᚛', '᚜' },
+//                    {'⁅', '⁆' },
+//                    {'⌈', '⌉' },
+//                    {'⌊', '⌋' },
+//                    {'❨', '❩' },
+//                    {'❪', '❫' },
+//                    {'❬', '❭' },
+//                    {'❮', '❯' },
+//                    {'❰', '❱' },
+//                    {'❲', '❳' },
+//                    {'❴', '❵' },
+//                    {'⟅', '⟆' },
+//                    {'⟦', '⟧' },
+//                    {'⟨', '⟩' },
+//                    {'⟪', '⟫' },
+//                    {'⟬', '⟭' },
+//                    {'⟮', '⟯' },
+//                    {'⦃', '⦄' },
+//                    {'⦅', '⦆' },
+//                    {'⦇', '⦈' },
+//                    {'⦉', '⦊' },
+//                    {'⦋', '⦌' },
+//                    {'⦍', '⦎' },
+//                    {'⦏', '⦐' },
+//                    {'⦑', '⦒' },
+//                    {'⦓', '⦔' },
+//                    {'⦕', '⦖' },
+//                    {'⦗', '⦘' },
+//                    {'⧘', '⧙' },
+//                    {'⧚', '⧛' },
+//                    {'⧼', '⧽' },
+//                    {'⸂', '⸃' },
+//                    {'⸄', '⸅' },
+//                    {'⸉', '⸊' },
+//                    {'⸌', '⸍' },
+//                    {'⸜', '⸝' },
+//                    {'⸠', '⸡' },
+//                    {'⸢', '⸣' },
+//                    {'⸤', '⸥' },
+//                    {'⸦', '⸧' },
+//                    {'⸨', '⸩' },
+//                    {'【', '】'},
+//                    {'〔', '〕' },
+//                    {'〖', '〗' },
+//                    {'〘', '〙' },
+//                    {'〚', '〛' }
+//                };
+//    }
+//}
